added find substring func and compare n characters

+++ stringfuncs/types.h
@@ -1,3 +1,3 @@
 #pragma once
 
-typedef unsigned int uint32;
+typedef unsigned int usize;
